requirements.txt
pandas
python-dateutil
pyyaml
matplotlib
pyarrow
fastparquet
streamlit
